renamed tests after service refactor
